Move the grouping logic to Mapbox layers' `filter` property
Now instead of having a single Source (and usually a single Layer) for
each feature group, I now have a single layer for each geometry type
(except for lines which have two: one for display and one for event
targets), and each layer filters both by geometry type and by the
currently visible layers. For example here's the point layer's filter:

```
[
  'all',
  ['==', ['get', 'layerType'], 'point'],
  ['in', ['get', 'featureGroup'], ['literal', Array.from(visibleGroups)]],
]
```

It checks both that the `layerType` is a point, and that the
`featureGroup` is in the `visibleGroups` set.

I like this!

It also solves the problem of z-indexes with polygon layers (because
now I can make sure they're always at the bottom), so I can comfortably
add `calmedTrafficArea` to the initial set of visible groups :)