Update the Overpass query's boundaries
I didn't know much about Overpass QL when I wrote this query. I still don't know a lot, but I do know that I can query for features inside areas instead of just bounding boxes! Specifically I can query inside the Tel Aviv and Central districts to get exactly the areas I care about!

This results in a pretty substantial increase in the size of the output (4MB compared to the previous 2MB).

But #1: I can change the output from `body` to `skel` because I don't really use any of the tags. This reduces the file size to 3.6MB (10% smaller!).

But #2: There are a lot of bike trails in forests in parks that aren't that useful for this app, so I can remove them:

- יער בן-שמן
- יער שהם
- שמורת טבע הדום שומרון
- שמורת טבע נחל שילה
- יער מודיעין (וחברים)
- נאות קדומים
- יער המגינים
- יער חורשים

Removing these gives a file size of 2.6MB, which is another 20% off the initial 4MB, and just 30% bigger than the original 2MB, which I think is worth it :)

Another important downside is that running this query now takes a lot longer, but that's fine because I have build minutes to spare.